[ALPS03425076] DISP:Revise ion/m4u buffer setting
[Detail] Because the type of mva_size is size_t,
the casting of uint32 might be worrisome on a
64-bit platform.

[Solution] Replace it by size_t.

MTK-Commit-Id: 52f9beff52cbc08a8fe68ba01a05bf27ccb938b9

CR-Id: ALPS03425076
Feature: Power Management

+++ drivers/misc/mediatek/video/mt6735/primary_display.c
@@ -2444,7 +2444,7 @@ static disp_internal_buffer_info *allocat_decouple_buffer(int size)
 		}
 		buf_info->handle = handle;
 		buf_info->mva = buffer_mva;
-		buf_info->size = (uint32_t)mva_size;
+		buf_info->size = mva_size;
 		buf_info->va = buffer_va;
 		buf_info->client = client;
 	} else {
